pyproject.toml
[tool.poetry]
name = "backend-set-cookie"
version = "0.1.0"
description = ""
authors = ["hamidreza <www.hamidreza121212@gmail.com>"]
readme = "README.md"
packages = [{include = "backend_set_cookie"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
